refactor:프로필페이지 스타일, 리팩토링
- 변경된 PostCard 가 결합될수있도록 코드와 스타일 변경
