+++ src/styles/components/post/_post-card.scss
@@ -49,9 +49,9 @@
 
   // ================================== gridlist =============================================================
   &--compact {
-    position: relative;
-    .post-card__overlay {
-    }
+    padding: 0;
+    margin:0;
+    background-color: black;
   }
 
   // ============================================== 해더 ===========================================================
