+++ src/styles/components/post/_postlist.scss
@@ -1,8 +1,6 @@
 .grid-posts {
   width: 100%;
-  border: 1px solid red;
-  max-width: 64rem;
   display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1rem;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap:1rem;
 }
