+++ src/styles/pages/_profile.scss
@@ -5,7 +5,11 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
+
   &__container {
+    width: 100%;
+    max-width: 1200px;
     padding: 2rem;
     display: flex;
     flex-direction: column;
@@ -14,14 +18,18 @@
   }
 
   &__header {
+    border: 1px solid v.$light-4;
+
+    background-color: v.$dark-3;
+    width: 100%;
     padding: 1rem;
     display: flex;
     margin-bottom: 2rem;
-    width: 100%;
+
     &-container {
+      width: 100%;
       padding: 1rem;
       display: flex;
-      flex: 1 1 0%;
       gap: 1rem;
     }
   }
